Extract FeedImageDataLoaderSpy into a shared scope to remove duplic…
…ation
